requirements.txt
pandas 
numpy
scikit-learn
flask
joblib
gunicorn
pandas-profiling
google-cloud-storage
pyyaml